Allow Monte Carlo Algo that uses heuristic for large search spaces.
When it gets under a certain size, it may need to try other words to
disambiguate a large fanout.

For example, if the remaining words are batch, catch, hatch, latch,
match, patch, and watch, then the expected turns left would be something
like 4. With a well-chosen word (maybe climb?) which it CANNOT be, the
algo would make the expected turns go to 2.<something>. A monte carlo
search solves this failure mode, which I've seen the algo get into.

The problem with a large search space for the monte carlo algorithm, is
that it's REALLY slow. I think I'm going to have it choose the top N
guesses based on the heuristic it's given. That'll need a small
refactoring, though.